sdk_ui: select synced styled system
Summary:
Add support for styled system for:

- `SelectSynced`

The following style props are allowed:

```
export type StyleProps = {|
    ...MaxWidthProps,
    ...MinWidthProps,
    ...WidthProps,
    ...FlexItemSetProps,
    ...PositionSetProps,
    ...MarginProps,
|};
```

A default width of `100%` is added.

This diff splits up the props from the helpers into separate `Select` and `SelectButtons` props since they don't completely overlap.
This diff adds examples for `SelectButtons` and `SelectButtonsSynced` in the UI playground, but doesn't add support for styled system for those components — that will come in a different diff.

Test Plan:
- Verify in UI playground

{F135655}

Reviewers: billylittlefield, s2tephen

Reviewed By: billylittlefield, s2tephen

Differential Revision: https://airtable.phacility.com/D11924