cli: Remove smoke test from release-it hooks.
Summary:
The smoke tests currently require stdin input (Press ENTER to continue), but when run through `release-it` it appears that stdin is not passed through.

In this diff we remove the smoke test from `release-it` hooks. They will now need to be run manually :(

Test Plan: No QA testing needed.

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D11944