examples: update records block
Summary:
Depends on D11859

Update records block! Like batch update, but only adds 1 to a certain numeric field, and uses selected records rather than a view.

My original idea was for a "calculator" block - you'd be able to select records and add/subtract/etc different values. But I decided that might be too complicated for an example, so I kept it to just adding 1 to fields. I think it'd be easy for users to modify it to do more operations.

Test Plan: No QA required

Reviewers: somehats, kasrak, maryrosecook

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D11739