sdk_ui: record card styled sytem
Summary:
- Adds styled system to `RecordCard`. The allowed style props are:
```type StyleProps = {|
    ...FlexItemSetProps,
    ...PositionSetProps,
    ...MarginProps,
|};```
- Converts as much of `RecordCard` towards `Box` / away from `baymax` as we can for now.
- Adds a `RecordCard` example in UI playground/

## Notes:
`RecordCard` already explicitly takes a `width` and `height` property, which are used internally for layout calculations. We'll continue to support these props, and for now will leave all the `dimensionsSet` style props out of `RecordCard`.

`RecordCard` exports using `withHooks`, so we need to "inject" the `useStyledSystem`-generated `className` in addition to any user-defined classNames to the component being wrapped. Technically, we're "updating" not "injecting" the `className` prop, so it's omitted from the list of `InjectedProps` in the type definition for flow appeasement purposes.

Test Plan: Verify that the `RecordCard` example behaves as expected in UI Playground.

Reviewers: jay-ransijn, s2tephen

Reviewed By: jay-ransijn, s2tephen

Differential Revision: https://airtable.phacility.com/D12002