sdk_ui: fix select emoji height, view picker prop types
Summary:
Maintains the height of `Select`, even when an emoji is present.

Also fixes prop types in `ViewPicker`.

Test Plan: In the `ViewPicker` example in the UI playground, switch between a view with an emoji in its name and one without. The height of the select should not change and there should be no errors in the console.

Reviewers: jay-ransijn, billylittlefield

Reviewed By: jay-ransijn

Differential Revision: https://airtable.phacility.com/D12160