sdk: convert to typescript
Summary:
This is the main typescript conversion diff. It was created from several distinct steps that we'll talk about in a moment, but can't really be broken down and merged piece-by-piece - this entire diff needs to be landed all in one go.

For the most part, the changes here were done by (a slightly modified version of) @calebmer's amazing typescript codemods for hyperbase. On top of that, several hundred type errors were manually fixed. Outside of the core code changes, this diff makes big changes to how code is transpiled and published, and completely replaces the documentation system used by the blocks SDK as our old one was incompatible with typescript.

Reviewing this diff itself is unlikely to be that valuable - it's too big, too much has changed. That said, some changes are in more need of review than others, and those have been pulled highlighted in the step-by-step diffs below.

We'll be dividing the review of this up into general areas to avoid too much review burden on a small number of people. Those areas and their reviewers are:

- UI kit - @jay-ransijn (blocking) and @s2tephen
- Model layer - @minicat (blocking) and @jbbakst
- Documentation - @EvanHahn (blocking)
- Surrounding tooling, package setup, utility types, etc - @calebmer (blocking)

More granular diffs to actually review, representing the step-by-step process that went into this diff:
1. D12218 - prepare to run the codemod (not much interesting here)
2. D12219 - run the codemod. 100% automated changes here.
3. **D12220 - fix syntax errors after running the codemod. Not much interesting here, but potentially a good one for @calebmer to look at**
4. D12221 - format code with prettier. 100% automated changes here. Probably not worth reviewing.
5. **D12222 - fix type errors. This diff has a lot of manual changes to code. It mostly touches types, but in some places tweaks the run-time semantics or exported interface of certain modules. This is the diff that could use the most careful review.**
6. **D12223 - new tooling. This updates the tooling around the sdk to support typescript, including eslint and build systems. most of the code-changes are about fixing lint issues.**
7. **D12224 - documentation system. This adds typedoc and generally completely reworks how documentation works for the SDK. One for @EvanHahn and the UI kit team to look at**
8. D12225 - bonus - a diff of the built files. This is how the actual generated output that gets run in your browser changed as a result of all this.

I'll be adding inline comments to each of these diffs to highlight areas that could use some explanation or greater attention.

Test Plan: I need to figure this bit out, but I'm planning on migrating one or two blocks to this version - possibly chart and translate. Suggestions welcome!

Reviewers: jay-ransijn, s2tephen, minicat!, jbbakst, EvanHahn, calebmer

Reviewed By: jay-ransijn, EvanHahn, calebmer

Subscribers: jbbakst, s2tephen, EvanHahn, jay-ransijn, minicat, calebmer

Differential Revision: https://airtable.phacility.com/D12235