sdk_ui: button revamp
Summary:
Revamps the `Button` component with `size` and `variant` props.

The `variant` prop replaces the `theme` API. This is a mapping to the previous API:

- default <= gray
- primary <= blue
- secondary <= transparent
- danger <= red

The `variant` prop will work through a `string` instead of accessing it on an enum.

This update will be a breaking change and we should provide guidance in the changelog.

Test Plan:
Verify storybook.
{F137955}

Reviewers: s2tephen, billylittlefield

Reviewed By: s2tephen, billylittlefield

Differential Revision: https://airtable.phacility.com/D12211