sdk_ui: icon names autocomplete
Summary:
Splits up all of the icon names in two 2 enums: `regularIconNames` and `microIconNames`.
Each will have a type `RegularIconName` and `MicroIconName`.
Each have a corresponding object that contains all the config: `regularIcons` and `microIcons`.
Each have a corresponding prop type: `regularIconPropType` and `microIconPropType`.

Finally, there is a `IconName` type and `icons` object that contains all of the above.

To make it seamless to keep Hyperbase and the UI kit the same as icons get added I created a utility (https://codesandbox.io/s/winter-sound-f4p2l) that takes the `iconConfig` and generates the contents of this file.

@billylittlefield if we decide to land `makeEnum` we might want to update this diff (and generator) to use `makeEnum` instead.

## Change in `createEnum` return type

This diff also change the return type of the `createEnum` utility to be `{[K in T]: K}` instead of `{[K in T]: T}`.

Test Plan:
- Verify storybook.
- Verify autocomplete in `Button` etc.
{F139731}

Reviewers: s2tephen, billylittlefield

Reviewed By: s2tephen

Subscribers: billylittlefield

Differential Revision: https://airtable.phacility.com/D12406