sdk: generate docs automatically on push to master
Summary:
This adds a GitHub action which will automatically build the SDK docs whenever we push a commit
to master. This means our documentation is always up-to-date, and if a developer builds the docs for
one of their changes, their diff will only include changes relevant to them.

Previously, we were building the docs in a pre-push hook. That also solved the problem, but at the
cost of running a fairly slow operation as a synchronous part of the development workflow. By moving
this to a GitHub Action, it happens asynchronously, so doesn't impact the dev workflow.

Depends on D12643

Test Plan: Tested manually on this branch

Reviewers: EvanHahn, s2tephen

Reviewed By: EvanHahn, s2tephen

Subscribers: philip.zeyliger, newyorkpizza

Differential Revision: https://airtable.phacility.com/D12648