sdk_ui: switch sizes & variants
Summary:
This diff accomplishes a number of tasks pertaining to the `Switch` (formerly `Toggle`) UI kit component:
- renames Toggle[Synced] to Switch[Synced]
- converts Switch[Synced] to functional components
- adds sizes and variants to Switch
- changes default styling of Switch to be full-width with a background color for easier use in forms

Because of the change to the default styling, `backgroundColor` has been added to the list of supported style props for easier overriding. Additionally, this diff automatically sets the `aria-labelledby` to an ID referencing the included label if a label node/text is provided. This can optionally be overridden by the `aria-labelledby` prop directly.

{F143937}

Test Plan: Verify the Switch / SwitchSynced components behave as expected in UI Playground and storybook.

Reviewers: jay-ransijn, s2tephen

Reviewed By: jay-ransijn

Differential Revision: https://airtable.phacility.com/D12979