cli: remove __gitignore from template
Summary:
Our template blocks include `__gitignore` files as `.gitignore` is ignored when publishing to npm.
When initializing a block using a template, we would copy `__gitignore` to `.gitignore` to make sure
the correct files were ignored for that block. That left the `__gitignore` file in the user's block
directory though, which is a little messy.

This diff switches that code over to move the file rather than copy it, meaning there's no `__gitignore`
left over in the user's final block dir.

Test Plan: Create a block with a template, ensure there's no `__gitignore` in the resulting folder.

Reviewers: maryrosecook, richsinn

Reviewed By: maryrosecook, richsinn

Subscribers: kasrak

Differential Revision: https://airtable.phacility.com/D13036