sdk: don't add icon margin to button when there is icon but no children
Summary:
Ran into this when migrating time tracker. There is a button with only an icon and no text which was asymmetrical because of this extra padding added when there is an icon.
Updated to only add the extra `Box` when there is children.

Also updated to make `children` an optional prop and instead `console.error` when there is no children and no aria-label (interim suggestion ahead of `IconButton` component / better support for this use case.)

(love the fact that `icon` is a prop now, btw! Makes it a lot cleaner)

Before:
{F147867}

After:
{F147869}
{F147868}

Storybook:
{F148394}

Test Plan: Tested in time tracker, and added to Storybook.

Reviewers: billylittlefield, jay-ransijn, s2tephen

Reviewed By: billylittlefield, s2tephen

Differential Revision: https://airtable.phacility.com/D13495