Docs: required Typedoc-related changes for docs-v2.5
Summary:
## Docs 2.5 background

Our latest iteration on SDK documentation is centered around a new application that lives in Hyperbase, and is a significantly different from past documentation approaches. The most notable difference is the shift from generated markdown to generated JSON output from Typedoc. Using the JSON output format gives us much more flexibility in customizing the content, look, & feel of the docs site, at the cost of requiring a larger initial engineering effort to build the surrounding application.

This diff makes necessary changes to Typedoc plugins and comments throughout the SDK codebase to support the new application.

## Changes made

List of all changes found in this diff:
1. New custom typedoc tags
2. Typedoc plugin changes
3. Typedoc script changes

#### 1. New tags
Four new tags have been added to support the new docs application. `@hook` and `@component` are simple tags to identify React hooks & components. These are used to provide a better heuristic when logic-splitting between a regular function/class and a React component function/class (ditto for hooks). Any new hooks or components should implement these tags in their typedoc to be supported.

The third tag is `@docsPath`, which tells the docs site where the documented reflection should live in the docs navigational tree. As of now, the tree is:
```
| /models
| -- /advanced
| /UI
| -- /components
| -- /hooks
| -- /utils
| -- /system
```
Reflections can live anywhere in this tree (not only leaf folders), and any new path will automatically be added to the site map, no plumbing required. The value for the tag should not be surrounded in quotes. Example of usage: `@docsPath models/Table`

The final new tag is `@groupWith`, which is used to signify when a reflection should be "grouped" with another reflection. As of now, this is only done for Synced components (eg, InputSynced). The value of the tag should be whatever the `@docsPath` value is for the reflection you want to group it with. We use this tag to render grouped reflections on one page with tab navigation to switch between them, and don't show a sidebar row for any reflection that contains a `@groupWith` tag. Think of it as a variant of whatever it's being grouped with. We want to still define a `@docsPath` so that it has a unique URL. For example, combine everything together and the tags on `InputSynced` look like this:
```
@docsPath UI/components/InputSynced
@groupWith UI/components/Input
@component
```

#### 2. Typedoc plugin changes
**CommentPlugin**
The comment plugin is enabled by default in Typedoc (no import needed). Unfortunately, it contains a bug where `comment` objects do not get set on functions with only one signature. This is a pretty common thing. We initially combatted this by converting all `function() {}` defintions to `const = () => {}` (artifacts of which are still in this diff), but ultimately this feels like an overly strict restriction that's annoying to enforce for future SDK developers. As a result, I forked the comment plugin, made a change that fixes the bug for our use case, and replace the default comment plugin with the forked version in `doc-generator/typedoc-plugin.js`. I also filed a bug with Typedoc [[ https://github.com/TypeStrong/typedoc/issues/1144 | here ]] - if I have some bandwidth I'll try to fix it in the opensource repo, for the sake of getting something out, the fork is our current solution.

**ExternalModulePlugin**
No changes were made to this plugin, but in theory we can get rid of it. The only reason I didn't was because I wasn't sure if there was some future value in keeping it that I'm not thinking of, and it doesn't cause any harm. If / when we do remove it, we can remove all the `@module` header tags on every file.

**HideInternalPlugin**
The `HideInternalPlugin` was hiding a bit too aggressively. Default reflection names for getters and setters were getting caught, so I added `__get` and `__set` to our name whitelist. Additionally, variable reflections nested under type literals were getting caught, but we actually want that information. We now unhide variables if their parent isn't an external module.

**SourcefileUrlPlugin**
This doesn't do anything for JSON output, and we can pretty easily piece together the sourcefile URL from the JSON anyway, so I removed this.

**MarkdownPlugin**
Removed for obvious reasons.

#### 3. Typedoc script changes
As described, we now use the `--json` option. The outputted filename is generated using the current package version. I removed the `out` option in `typedoc.js` to prevent the default HTML output in addition to the JSON

Test Plan: No QA needed for the SDK, as no source code is actually touched. We will QA the Typedoc JSON output as a result of QAing the entire blocks docs site in hyperbase (diff coming soon).

Reviewers: jay-ransijn, s2tephen, somehats

Reviewed By: s2tephen, somehats

Differential Revision: https://airtable.phacility.com/D13186