cli: fix console.log in backend functions
Summary:
Why
===
This is a forward port of D13915 (from `hyperbase` repo). With the upgrade
to Node v12, this started failing; uncertain why it was not an issue
with Node v8. It is also not an issue when running locally with `localhost` (even
on Node v12 running locally).

What
====
Forward port fix from D13915

Test Plan: Tested manually

Reviewers: kasrak, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D14020