Handle github templates
Summary:
Changes cli as part of https://airtable.com/tblirJUiB3DryM26j/viweryBDQNKvrnKpA/recMLBngcBuhwtPjA?blocks=bipdCpWkCd7sxO47r
I'm just reusing the template parameter based on my discussion with Kasra, this has the benefit of not needing to handle the case
where we receive two templates if we add a new parameter.

Test Plan: Run the tests. Init a block with a github repo as the template parameter

Reviewers: kasrak, minicat

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D14118