sdk_ui: view picker example
Summary:
Adds the ViewPicker example, which is really just using a modified Select under the hood to avoid dealing with stubbing tables/views.

This also changes the expected format of the `option.options` expected in `<Example>`. Previously, we accepted an object, but only were using the keys on the object. If we want to follow this paradigm, it probably makes more sense to pass as an array. I took it one step further by changing the format to an array of objects with `value` and `label`, which lets us set a different value in the control than what the underlying value is. I found myself needing this when adding `allowedTypes` option to the `ViewPicker`. I can split this into a separate diff if we want to go forward with the paradigm; I just first wanted to collect opinions.

ViewPicker example:
{F153749}

Test Plan: QA can skip

Reviewers: jay-ransijn

Reviewed By: jay-ransijn

Differential Revision: https://airtable.phacility.com/D14217