sdk: fix crashes caused by column type changes
Summary:
Fixes https://rollbar.com/Airtable/Airtable-Blocks/items/20163/

### Issue
I was able to reproduce the error from the above rollbar by adding a sort to a Record List block, and then changing the field type of the field being sorted. Specifically, going from a Single Select => Single Line Text would trigger it (going backwards, however, did not). Digging through the stack trace, it was evident that the `BlockVisList` was stale - we were trying to set a text value like "foo", but the column type provider was still geared for Single Select.

### Technical background
This appears to be an order-of-operations issue. In `__applyModelChanges`, we pass a big object of changes and then parse our way through the object, triggering any watch callbacks along the way. Changing the field type specifically results in a number of model changes:
- `name`, `type`, `typeOptions` for the field being updated
- `cellValue` for every record in the table

In `Table.__triggerOnChangeForDirtyPaths`, we currently trigger updates to the record store first, then doing everything else (ie, the record updates happen before the field updates). So, the record tries to update using a stale column provider, because the field update is where we trigger `replaceVisList` to update the column providers.

### Fix
The issue is resolved by doing the field updates **before** the record store updates. This did make me wonder, however, whether there are other "dependency chains" like this that aren't being respected. We could set-up a required order of operations for each model to formalize how this works. For now, I just shuffled the code around and added a comment.

Test Plan:
1) Create a table with a single select field, add some options and fill in some records with these options
2) Add a record list block to a table
3) Add a sort to the block settings for the single select field
4) Change the single select field into a single line text field
5) Change it back to a single select field

No crashing should occur

Reviewers: kasrak, somehats

Reviewed By: kasrak

Differential Revision: https://airtable.phacility.com/D14298