sdk: better errors on useWatchable
Summary:
Feedback item from Blocksathon - it's confusing that `useWatchable` will do nothing if there are no watchable keys.

This diff adds an error in that case, and also updates the typing of `useWatchable` to reflect that `undefined` is not a valid value. (Users should use `null` if they want to watch a key conditionally).

Also, I looked into printing the list of watchable keys alongside the error message, but it seemed like I'd need to hardcode a function that mapped the model being watched to the `WatchableXKeys` for that model, and that didn't seem super high impact.

Tested with a test block and added a unit test.
{F155164}

Test Plan: No QA required

Reviewers: somehats, kasrak

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D14197