cli: Make smoke test example block always uses the latest SDK version
Summary:
Currently, the example backend block uses a fixed SDK version
that is older than the latest release. It seems desirable to ensure
that the latest SDK version is used during the test to avoid false
negatives - remembering that we're testing the CLI here, and not the
SDK itself. The mechanics of this are trivial as we have a working
`latest` tag on the package already which yarn can substitute for us.

I've also removed the checked in package-lock.json. As far as I can
tell, this is not used today, and is definitely undesirable. You
can't generate a lock file until after the substitution is done,
and this happens as part of the actual block build during the test.
So it's a lot clearer to not have the file at all, than to suggest
it does anything by being present.

Test Plan:
Ran the smoke test to successful completion. Smoke test will
fail if the SDK version is not substituted.

Reviewers: richsinn, fuzziqersoftware

Reviewed By: richsinn

Subscribers: fuzziqersoftware

Differential Revision: https://airtable.phacility.com/D14534