cli: make github validation more permissive
Summary:
When talking to Stephen on the hyperbase part of this feature we had some back and forth on which url's this should allow.
I think it's better to make this a less stringent check so that we don't have to make as many updates in the future if we
change the web logic and also allow power users to have more options when using the cli directly.

Test Plan: I ran the unit tests and initialized a block with blockDev both with an npm and github template

Reviewers: kasrak, minicat

Reviewed By: kasrak

Differential Revision: https://airtable.phacility.com/D14560