cli: Remove non-essential metadata from package.json during `init`
Summary:
The init step takes a pre-existing template and creates a skeleton
for development of a new block. When the template comes from npm,
the package.json will have a lot of irrelevant metadata in it that
makes the file harder for the developer to work with, and also
includes the identity metadata for the template itself, which won't
be what anyone wants to carry forward.

So, as we believe we know which fields we actually want to retain,
let's apply a whitelist to create the final package.json.

Test Plan:
* Unit tests
* Manually inspected package.json after initing from existing
  templates

Reviewers: richsinn, somehats, fuzziqersoftware

Reviewed By: richsinn, somehats

Differential Revision: https://airtable.phacility.com/D14576