sdk: Push hooks more strongly in `QueryResult` docs
Summary:
A user unaware of hooks and perusing the docs for `Table` or `View`, looking to get records or view metadata, would scroll down to `selectRecords()` or `selectViewMetadata()` and then examine the return type `RecordQueryResult` or `ViewMetadataQueryResult`.

Edited their docstrings to recommend using hooks more strongly.

Test Plan: N/A

Reviewers: somehats, billylittlefield

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D14677