sdk: add details and example to Cursor doc
Summary:
As per discussion on D14693, add some more details to `Cursor` as we're not elaborating on it / linking to examples from the read guide.

Depends on D14532 (Kasra's `selectedFieldIds` diff) since I figured it's easier to write the text once instead of updating it after that lands.

Test Plan: No QA needed

Reviewers: billylittlefield, kasrak

Reviewed By: billylittlefield

Differential Revision: https://airtable.phacility.com/D14732