docs: cleanup & refactor @returns tags
Summary:
Currently, we don't do anything with the value of the `@returns` tag in the documentation site. The type information is already conveyed by the reflection, but there's actually quite a bit of extraneous information encoded in these comment lines. Sometimes, the lines are just copy-pasted into the typedoc comment as well, but other times it contains unique information that we want to expose on the docs site.

One solution would be to add support for rendering the `@returns` tag information in the docs site. The other is to "hoist" this content into the main comment section, which we already render for all reflection types. I opted for the latter, because there's some ambiguity about where we would render the `@returns` stuff, and we already have a mixed convention of sometimes describing the returned value in the comment and other times not. This diff standardizes towards exclusively using the comment.

This also fixes a few minor docs issues that I noticed when going through.

Test Plan: No qa needed

Reviewers: jay-ransijn, minicat, s2tephen

Reviewed By: jay-ransijn

Differential Revision: https://airtable.phacility.com/D14684