sdk: standardize FieldType and ViewType enums
Summary:
Depends on D14759 in hyperbase.

We currently have nonstandard treatment of the field type and view type enums, exporting them as `fieldTypes` and `viewTypes` from `@airtable/blocks/models`. This diff switches the named exports to the more canonical TS naming of `FieldType` and `ViewType`, and consolidates usage in the SDK to all reference these enums.

It also does an overhaul of the comments on `FieldType` enum members so they'll be displayed better in the enum table added to hyperbase in D14759. Now, each field type member will include:
- field options
- cell read value
- cell write value

Screenshot:
{F157804}

Test Plan: No qa needed

Reviewers: somehats, minicat, jay-ransijn

Reviewed By: minicat, jay-ransijn

Differential Revision: https://airtable.phacility.com/D14758