cli: Add blocks cli user-agent to /registerBlockInstallationMetadata response
Summary:
This allows us to display an error message in the block installation iframe if the blocks cli
version is unsupported.
I had to put this in the response body instead of the headers because cors causes the headers
to be stripped.

Test Plan: Tested that blockDev run still works

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D15029