examples: dont crash when new table is created but cursor.activeTableId is not null
Summary: Noticed this bug while testing something using table structure. When creating a new table, `cursor.activeTableId` is not null but the block hasn't received the table creation yet, so `base.getTableById` crashes. Updated youtube preview and table structure to be resilient to this, and will add some notes in the cursor docs with these conditions in a separate diff.

Test Plan: Tested the example blocks

Reviewers: benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D15197