sdk: update SettingsButton docs
Summary:
This came up during the Miro review - they were using `useSettingsButton` but no-opping in the onclick depending on permissions. I asked them to update to the pattern shown in  the updated example below to conditionally show the button instead.

Thought it'd be useful to document this as well as make the example use `useWatchable` instead of `.watch()` (which was incorrectly missing the `.unwatch()` before).

Test Plan: No QA required

Reviewers: billylittlefield

Reviewed By: billylittlefield

Differential Revision: https://airtable.phacility.com/D15377