mirror a redacted repo to public git rather than publishing the whole thing
Summary:
This adds an internal tool, `git-mirror`. This command is used for syncing a partial copy of the monorepo whilst scrubbing git history. It relies on a `git-mirror.json` config file which defines:
- The committer name/email for publish commits (`authorName`/`authorEmail`)
- The git remote to mirror to (`remote`)
- An allow-list of files which should be included in every publish commit (`global`)
- A list of publishing scopes defining a subset of files that can be published in each commit (`scope`)

The `git-mirror` command has two subcommands: `create` and `sync`. The `create` command uses the existing history to create a repo with only publish commits and no progress commits. The `sync` command takes a tag name and a scope name and creates a commit in the remote repository publishing just the files defined in that scope in a single commit with no details.

Output of the `create` command can be found here: https://github.com/Hyperbase/D14709-public-airtable-blocks-mirror-example (note the lack of a cli package and the redacted git history)

The `sync` command will be used as part of our standard `yarn release` process for packages from now on.

Test Plan: -

Reviewers: richsinn, kasrak, somehats

Reviewed By: richsinn

Subscribers: kasrak, billylittlefield

Differential Revision: https://airtable.phacility.com/D14709