git-mirror: remove standalone comments
Summary:
Depends on D14709

This adds a step to the `git-mirror` workflow for copying files that strips line comments (double slash comments) from javascript files. This would give us a means to separate public-facing comments from internal comments in the SDK codebase, as we could continue to use block comments to convey information. There's an escape hatch for line comments as well, which can be marked as protected with a ! immediately following the double slash.

This forks a library (https://github.com/jonschlinkert/strip-comments) into an internal tool, and makes a few changes:
 - fixes a bug where line comments inside of block comments were being stripped
 - fixes a bug where the comment-parsing regex got confused by empty strings
 - if the line comment was on its own line, it removes the whole line (newline included)
 - whitelists "@flow" and "eslint-disable" comments and does not remove them if the comment contains that text

Test Plan: Audit the output of this process compared to master: https://github.com/Hyperbase/blocks-sdk/pull/7/files

Reviewers: kasrak, richsinn, benrazon

Reviewed By: richsinn

Subscribers: somehats

Differential Revision: https://airtable.phacility.com/D14842