blocks cli: switch request dependency to postman-request
Summary:
https://airtable.com/tbl3f9cb8b2f15ab8/viw4B6ukenBk6WMFF/rec09IJoFnoKp88mg?blocks=hide
The maintainers of the request npm package recently deprecated it. Some people on that issue recommended
postman-request as a drop in replacement that was forked from request and is slightly more maintained.
This seemed like the simplest fix to remove the deprecation warning.

Test Plan: I ran the smokeTest and installed the cli from npm pack and saw that there is no longer a deprecation warning

Reviewers: richsinn, kasrak

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D15571