name_quiz: fix height
Summary:
I was hoping that the previous diff would also have addressed the height issue in Desktop, but somehow still needs `height: 100%;`.
There is not a great way to test Desktop without releasing so I hope this should fix it (I tried it in the inspector).

Test Plan: - Verify pictures show up in Desktop

Reviewers: benrazon, s2tephen

Reviewed By: s2tephen

Differential Revision: https://airtable.phacility.com/D15629