Base schema: don't consider deleted table when placing tables
Summary:
https://airtable.quip.com/ZA2qAQoLb4bD/March-2020-new-blocks-feedback
New tables were being displayed very far to the right. This was because we weren't
removing deleted tables from the table coord config.

Test Plan:
Add a table (observe its directly to the right of the right-most table)
Move the table very far to the right
Delete the table
Add a new table (observe its directly to the right of the right-most table)

Reviewers: minicat, billylittlefield

Reviewed By: billylittlefield

Subscribers: billylittlefield

Differential Revision: https://airtable.phacility.com/D15756