flowchart: disable exporting when settings are invalid
Summary:
Bug: https://airtable.com/tbl3f9cb8b2f15ab8/viw40mH4fGClHsR8K/recsjv76ztrasVh4F?blocks=hide

Currently, if you try to export as SVG when the settings are invalid, you can crash the block.

This fix disables the button when the settings are invalid to prevent that error.

Test Plan: Verify that you can only export as PNG/SVG when the settings are complete (ie, a flowchart is present in the block).

Reviewers: benrazon, kasrak, jay-ransijn

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D15803