url_preview: always show supported URL button
Summary:
- Always show button for supported URLs on empty state
- Change copy from "selected record" to "selected cell"
- Simplify function to get url preview and change name
- Refactor how content + dialog is rendered

I ended up refactoring a bit, and I am no fully convinced this is the best way to refactor.
In this current setup the dialog will on be closed when the user closes it, and not when you select a different cell.
This also allows it be easy to share the code for the dialog for the 2 empty states.
The empty states could be abstracted more, but I am afraid that comes with more complexity.

{F166812}
{F166813}
{F166823}

Test Plan:
- Open block
- Change table (or refresh)
- Verify empty state says "Select a cell to see a preview"
- Verify "View supported URLs" text button appears
- Click button
- Verify dialog appears with supported URLs

Reviewers: kasrak, billylittlefield, benrazon

Reviewed By: billylittlefield

Differential Revision: https://airtable.phacility.com/D16006