Make npm access public in release-it config section
Summary:
I think this is why releasing a new cli version caused it to become private.
I changed the release-it config for all the npm packages we switched to public

Test Plan:
Not testing since we don't have anything to release for now.
Let's keep an eye out in the next release

Reviewers: richsinn, billylittlefield

Reviewed By: richsinn, billylittlefield

Differential Revision: https://airtable.phacility.com/D16019