Flowchart: escape double quotes
Summary:
https://airtable.com/tbl3f9cb8b2f15ab8/viw7P84NHWqHTImvH?blocks=hide
Some users have " in their record values which causes the flowchart block to crash.
https://www.graphviz.org/doc/info/lang.html mentions that the only escapes character is double-quote,
so I think this is both necessary and sufficient as more general escaping leads to escaped characters
appearing in the label.

Test Plan:
Include a quote in a visualized record's primary field. Observe that it is displayed instead
of crashing

Reviewers: s2tephen, billylittlefield

Reviewed By: s2tephen

Differential Revision: https://airtable.phacility.com/D16136