Flowchart: add max record limit
Summary:
https://airtable.com/tblJvlqwVOgRLmgdT/recYmjVrqqPlISA3R/comeUlvuchzHN6cNT?fromNotification=notMkcwJKG4ri3w7Y
Looks like this is a known issue with vis.js on certain browser/os combos. There isn't a clear max number of
nodes as it does depend on the frame sizes and graph topology, but in my testing I saw this with 280-320 records
so hopefully 200 is both sufficiently safe and sufficiently large.

Test Plan:
Set up the block to use a view with more than 300 records. See that it now shows a message instead of
crashing.

Reviewers: s2tephen, billylittlefield

Reviewed By: s2tephen

Subscribers: maryrosecook

Differential Revision: https://airtable.phacility.com/D16138