cli: check envelope size during dev creds decrypt
Summary:
Depends on D16225

This is for first-party blocks with dev credentials that use the CLI tool for block building.
This is modeled after D16128, except one difference is that we'll throw an Error to halt execution instead of returning null because this is indicative of a biger issue.

Test Plan: No QA Needed

Reviewers: cakoose

Reviewed By: cakoose

Differential Revision: https://airtable.phacility.com/D16315