cli: allow non-isolated release builds
Summary:
Normally, `block release` creates a build in complete isolation from the existing block dir - it
copies all the files over to a new temp dir, runs `npm ci`, and then builds more or less as
normal. This strategy is designed to replicate the legacy block builder service.

This has a few drawbacks:
- No code-sharing between blocks in sibling directories
- No code-sharing using relative link/file-based `package.json` dependencies
- `yarn` users might get dependencies that differ from `yarn.lock`
- Different process from `block run` might be surprising

This diff adds a new option to `block release`: `--disable-isolated-build`. This option makes
release builds behave much more similarly to development builds - we build in the same folder as is
being developed in, and don't run `npm install`.

In the future, it could make sense to make this behavior the default, but for now this option is
motivated by a desire to move the scripting block into hyperbase and have it share some dependencies
with hyperbase.

Test Plan:
Builds without `--disable-isolated-build` are unchanged
Builds with `--disable-isolated-build` skip npm install etc. but still work successfully

Reviewers: richsinn

Reviewed By: richsinn

Subscribers: jbbakst, jichu4n

Differential Revision: https://airtable.phacility.com/D16499